feat: add JITP support
This commit adds a new service jitp-watcher which is
enabled by the config content interface connect hook
if the interface is connected, but no config file is
found. The jipt-watcher service is a shell script
that uses inotifywait to watch for a config.yaml file
to be copied or moved to watched directory, and once
found, automatically sets the greengrass-config option
and starts/enables the Nucleus.

Fixes: #19

Signed-off-by: Tony Espy <espy@canonical.com>