Add basic JITP support (#43)
* feat: add JITP support

This commit adds a new service jitp-watcher which is
enabled by the config content interface connect hook
if the interface is connected, but no config file is
found. The jipt-watcher service is a shell script
that uses inotifywait to watch for a config.yaml file
to be copied or moved to watched directory, and once
found, automatically sets the greengrass-config option
and starts/enables the Nucleus.

Fixes: #19

Signed-off-by: Tony Espy <espy@canonical.com>

* fix: make example config content writable

Use $SNAP_DATA for shared-files so that JITP
can be simulated by removing config.yaml before
building this snap, which allows it to be copied
into place at a later time, triggering the jitp
watcher in the greengrass snap to set the config
option.

Signed-off-by: Tony Espy <espy@canonical.com>

* fix: update snap hooks install checks

This commit updates the install logic in
the snap hooks to check for existence of
the IPC socket vs. GG_ROOT, as the latter
can be created when the config content
interface is connected.

Signed-off-by: Tony Espy <espy@canonical.com>

* fix: minor snap hooks cleanup

 - ensure hooks all pass shellcheck
 - always use $SNAP_NAME vs. hard-coded name

Signed-off-by: Tony Espy <espy@canonical.com>